it works

+++ src/LineSegment.java
@@ -1,13 +1,13 @@
 
 public class LineSegment {
 	
-	int x1;
-	int y1;
+	float x1;
+	float y1;
 	
-	int x2;
-	int y2;
+	float x2;
+	float y2;
 	
-	LineSegment(int x1,int y1, int x2, int y2) {
+	LineSegment(float x1,float y1, float x2, float y2) {
 		this.x1 = x1;
 		this.y1 = y1;
 		this.x2 = x2;
